+++ Assets/OxGFrame/NetFrame/Scripts/Runtime/Core/Interface/NetOption/NetOption.cs
@@ -1,0 +1,12 @@
+namespace OxGFrame.NetFrame
+{
+    public abstract class NetOption
+    {
+        public int autoReconnectCount { get; set; }
+
+        public NetOption(int autoReconnectCount = -1)
+        {
+            this.autoReconnectCount = autoReconnectCount;
+        }
+    }
+}