+++ Assets/OxGFrame/NetFrame/Scripts/Runtime/Core/Interface/NetTip/INetTips.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace OxGFrame.NetFrame
+{
+    public interface INetTips
+    {
+        void OnConnecting();
+
+        void OnConnected(object status);
+
+        void OnConnectionError(string msg);
+
+        void OnDisconnected(object status);
+
+        void OnReconnecting();
+    }
+}